+++ server/src/controllers/user.controller.ts
@@ -1,13 +1,14 @@
-import { RequestHandler } from "express";
+import { Request, RequestHandler } from "express";
 import asyncHandler = require("express-async-handler");
+import { matchedData } from "express-validator";
 import { createUser, getUser, getUsers, deleteUser } from "../services/user";
 
-import { UserId } from "../interfaces/user";
+import { CreateUser, UserIdParams } from "../interfaces/user";
 import validateCreateUser from "../validators/createUser.validator";
 
 interface userController {
   createUser: RequestHandler;
-  getUser: RequestHandler<UserId>;
+  getUser: RequestHandler;
   getUsers: RequestHandler;
 }
 
@@ -15,13 +16,16 @@ const userController = {
   createUser: [
     validateCreateUser(),
     asyncHandler(async (req, res) => {
-      const user = await createUser(req.body);
+      const newUser = matchedData<CreateUser>(req, { onlyValidData: true });
+      const user = await createUser(newUser);
       res.json(user);
     }),
   ],
   getUser: asyncHandler(async (req, res) => {
-    const { id } = req.params;
-    const user = await getUser(id);
+    const { userId } = matchedData<UserIdParams>(req, {
+      onlyValidData: true,
+    });
+    const user = await getUser(userId);
     res.json(user);
   }),
   getUsers: asyncHandler(async (req, res) => {
@@ -29,8 +33,10 @@ const userController = {
     res.json(users);
   }),
   deleteUser: asyncHandler(async (req, res) => {
-    const { id } = req.params;
-    const user = await deleteUser(id);
+    const { userId } = matchedData<UserIdParams>(req, {
+      onlyValidData: true,
+    });
+    const user = await deleteUser(userId);
     res.json(user);
   }),
 };