+++ server/src/interfaces/blog.ts
@@ -1,0 +1,5 @@
+export interface Blog {
+  id: string;
+  authorId: string;
+  content: string;
+}