+++ server/src/interfaces/comment.ts
@@ -1,0 +1,6 @@
+import { Blog } from "./blog";
+import { CreatedAt } from "./post";
+
+export interface Comment extends Blog, CreatedAt {
+  postId: string;
+}